+++ tests/tests.py
@@ -115,18 +115,6 @@ class MyPermissionsTests(JiraTestCase):
         self.assertGreaterEqual(len(perms["permissions"]), 10)
 
 
-@not_on_custom_jira_instance
-class ResolutionTests(JiraTestCase):
-    def test_resolutions(self):
-        resolutions = self.jira.resolutions()
-        self.assertGreaterEqual(len(resolutions), 1)
-
-    def test_resolution(self):
-        resolution = self.jira.resolution("2")
-        self.assertEqual(resolution.id, "2")
-        self.assertEqual(resolution.name, "Won't Fix")
-
-
 class SearchTests(JiraTestCase):
     def setUp(self):
         JiraTestCase.setUp(self)
